Move parser to args package and its creation to a factory module
